Upgrade to LLVM 21 (#71)
This may look like a troll PR, but `cargo test` works for me with LLVM
21.1.1 on Debian stable.

Well, I have only tested in a fork that uses static linking instead of
`-lMLIR` and I don't have a `libMLIR.so` on hand, so please test this
yourself. But it should work. (The hack to do static linking of MLIR
libs [[1][1],[2][2]] is unrelated and almost certainly too ugly to be
merged.)

Related: mlir-rs/tblgen-rs#28

[1]:
https://github.com/gt-tinker/qwerty/blob/2abca7773b5e8dc681ff6be05b3b9b73d7d59c1d/CMakeLists.txt#L122-L157
[2]:
https://github.com/gt-tinker/qwerty-mlir-sys/blob/9174f41e78d82517d5394ce4d61a12ecd2440ac6/build.rs#L169

---------

Co-authored-by: McCoy Becker <mccoybecker@gmail.com>